pyproject.toml
[tool.pdm]

[project]
name = "nonebot_plugin_orangedice"
version = "0.3.4"
description = "A COC dice plugin for nonebot2"
authors = [
    {name = "BigOrangeQWQ", email = "2284086963@qq.com"},
]
dependencies = [
    "nonebot2>=2.0.0b5",
    "nonebot-adapter-onebot>=2.1.3",
    "pydantic>=1.9.2",
    "sqlmodel>=0.0.8",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0"]
build-backend = "pdm.pep517.api"
